+++ src/main/java/ladder/domain/Player.java
@@ -2,21 +2,13 @@ package ladder.domain;
 
 public class Player {
 
-    private static final int MAX_NAME_LENGTH = 5;
-    private final String name;
+    private final Name name;
 
     public Player(String name) {
-        validateName(name);
-        this.name = name;
-    }
-
-    private void validateName(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException("참여자의 이름은 최대 " + MAX_NAME_LENGTH + "글자를 넘을 수 없습니다.");
-        }
+        this.name = new Name(name);
     }
 
     public String getName() {
-        return name;
+        return name.getValue();
     }
 }
