+++ src/test/java/ladder/domain/NameTest.java
@@ -1,0 +1,15 @@
+package ladder.domain;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import org.junit.jupiter.api.Test;
+
+class NameTest {
+
+    @Test
+    void 참여자의_이름은_5자를_초과할_수_없다() {
+        assertThatThrownBy(() -> {
+            new Name("5글자넘어가는이름");
+        }).isInstanceOf(IllegalArgumentException.class);
+    }
+}
